sync form.cpp

+++ Giun/wuucontest/server.cpp
@@ -6,7 +6,7 @@
 #define opt ios_base::sync_with_stdio(0); cin.tie(0)
 #define er(variable) cerr << variable << " "
 #define erl(variable) cerr << variable << "\n"
-#define loop(variable, valuegan, valuebehonbang) for(long long variable = valuegan; variable <= valuebehonbang; variable++)
+#define loop(variable, valuegan, valuebehonbang) for(ll variable = valuegan; variable <= valuebehonbang; variable++)
 #define loopd(variable, valuegan, valuelonhonbang) for(long long variable = valuegan; variable >= valuelonhonbang; variable--)
 #define loopdk(variable, valuegan, conditions, thaydoi) for(long long variable = valuegan; conditions; thaydoi) 
 #define pp(type, type1) pair<type, type1>
